requirements.txt
setuptools
flask
python-dotenv
flask-cors
gunicorn
paperscraper
nltk
openai
pymupdf==1.25.2
pymupdf4llm
langchain-openai
langchain-core
langchain-ollama
langgraph
typing-extensions
pydantic
duckduckgo-search
tavily-python 